requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pygithub==2.1.1
python-dotenv==1.0.0
sqlalchemy==2.0.23
anthropic==0.21.0
tree-sitter==0.23.2
tree-sitter-python==0.23.6
semgrep==1.68.0
celery==5.3.6
redis==5.0.1
pydantic==2.5.0
pydantic-settings==2.0.3
requests==2.31.0
psycopg2-binary==2.9.10